- BREAKING CHANGE: iterator_facade::pointer now corresponds to the actual result of iterator_facade::operator-> rather than Value*. This required an adjustment to a test. - The logic for determining the result of iterator_facade::operator[] has been factored out into a separate detail header in preparation for its potential use in iterator_range to avoid iterator_range::operator[] from returning a reference to a temporary.
git-svn-id: http://svn.boost.org/svn/boost/trunk@80901 b8fc166d-592f-0410-95f2-cb63ce0dd405

+++ libs/iterator/test/indirect_iter_member_types.cpp
@@ -82,7 +82,7 @@ int main()
     typedef boost::indirect_iterator<char**, int, std::random_access_iterator_tag, long&, short> Iter;
     STATIC_ASSERT_SAME(Iter::value_type, int);
     STATIC_ASSERT_SAME(Iter::reference, long&);  
-    STATIC_ASSERT_SAME(Iter::pointer, int*);  
+    STATIC_ASSERT_SAME(Iter::pointer, long*);  
     STATIC_ASSERT_SAME(Iter::difference_type, short);  
   }
   return 0;
